requirements.txt
pyyaml
joblib
dagshub==0.3.8.post2
dvc==3.49.0
dvc-s3==3.1.0
joblib==1.3.2
mlflow==2.8.0
pandas==2.1.2
scikit-learn==1.3.2